--- src/com/notleh/services/PreparePlayers.java
@@ -1,17 +1,0 @@
-package com.notleh.services;
-
-import com.notleh.entities.Hand;
-import com.notleh.entities.Player;
-
-import java.io.FileNotFoundException;
-import java.util.ArrayList;
-
-import static com.notleh.enums.EnumHandScores.ZER0;
-import static com.notleh.services.Game.game;
-import static com.notleh.services.InputSeperateSort.createCardArray;
-
-public class PreparePlayers {
-
-
-
-}
